requirements.txt
asgiref==3.3.1
autopep8==1.5.6
Django==3.1.7
django-concurrency==2.3
django-filter==2.4.0
djangorestframework==3.12.2
gunicorn==20.0.4
pycodestyle==2.7.0
pytz==2021.1
rope==0.18.0
sqlparse==0.4.1
toml==0.10.2
